requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
openai==1.6.1
python-dotenv==1.0.0
pytz==2023.3
plotly==5.18.0
